Move object adding to view in toolbar appear
